Use doubles for sim

+++ src/wave_sim.hpp
@@ -9,29 +9,33 @@
 class WaveSim {
 public:
     WaveSim(
-        const int size, const float wave_speed, const float grid_spacing, const float timestep, const float loss = 1.0f)
+        const int size,
+        const double wave_speed,
+        const double grid_spacing,
+        const double timestep,
+        const double loss = 1.0)
         : c_size(size)
         , c_wave_speed(wave_speed)
         , c_grid_spacing(grid_spacing)
         , c_timestep(timestep)
         , c_loss(loss)
-        , m_buffer_past(c_size * c_size, 0.0f)
-        , m_buffer_present(c_size * c_size, 0.0f)
-        , m_buffer_future(c_size * c_size, 0.0f)
+        , m_buffer_past(c_size * c_size, 0.0)
+        , m_buffer_present(c_size * c_size, 0.0)
+        , m_buffer_future(c_size * c_size, 0.0)
     {
     }
 
-    void set_at(const Vector2i pos, const float value)
+    void set_at(const Vector2i pos, const double value)
     {
         m_buffer_present[pos_to_idx(pos)] = value;
     }
 
-    [[nodiscard]] float value_at(const Vector2i pos) const
+    [[nodiscard]] double value_at(const Vector2i pos) const
     {
         return value_at_idx(pos_to_idx(pos));
     }
 
-    [[nodiscard]] float value_at_idx(const size_t idx) const
+    [[nodiscard]] double value_at_idx(const size_t idx) const
     {
         return m_buffer_present[idx];
     }
@@ -67,35 +71,35 @@ private:
         return pos.x >= 0 && pos.x < c_size && pos.y >= 0 && pos.y < c_size;
     }
 
-    [[nodiscard]] float spatial_derivative_at_idx(const size_t idx) const
+    [[nodiscard]] double spatial_derivative_at_idx(const size_t idx) const
     {
 
         constexpr std::array<Vector2i, 4> neighbors { { { -1, 0 }, { 1, 0 }, { 0, -1 }, { 0, 1 } } };
-        float neighbor_sum = 0.0f;
+        double neighbor_sum = 0.0;
         const auto [x, y] = idx_to_pos(idx);
         for (const auto& [n_x, n_y] : neighbors) {
             if (const Vector2i neighbor { x + n_x, y + n_y }; in_bounds(neighbor)) {
                 neighbor_sum += m_buffer_present[pos_to_idx(neighbor)];
             }
         }
-        const float numerator = neighbor_sum - 4.0f * m_buffer_present[idx];
-        const float denominator = c_grid_spacing * c_grid_spacing;
+        const double numerator = neighbor_sum - 4.0 * m_buffer_present[idx];
+        const double denominator = c_grid_spacing * c_grid_spacing;
         return numerator / denominator;
     }
 
-    [[nodiscard]] float future_at_idx(const size_t idx) const
+    [[nodiscard]] double future_at_idx(const size_t idx) const
     {
         return c_wave_speed * c_wave_speed * spatial_derivative_at_idx(idx) * c_timestep * c_timestep
-            - m_buffer_past[idx] + 2.0f * m_buffer_present[idx];
+            - m_buffer_past[idx] + 2.0 * m_buffer_present[idx];
     }
 
     const int c_size;
-    const float c_wave_speed;
-    const float c_grid_spacing;
-    const float c_timestep;
-    const float c_loss;
-    std::vector<float> m_buffer_past;
-    std::vector<float> m_buffer_present;
-    std::vector<float> m_buffer_future;
+    const double c_wave_speed;
+    const double c_grid_spacing;
+    const double c_timestep;
+    const double c_loss;
+    std::vector<double> m_buffer_past;
+    std::vector<double> m_buffer_present;
+    std::vector<double> m_buffer_future;
     BS::thread_pool m_thread_pool;
 };
